+++ tests/test_dispatchers.py
@@ -188,7 +188,7 @@ class TestDispatcherAcceptableHandlers:
         def handler(*_):
             pass
 
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
 
         dispatcher.add_handler(collector.handlers[re.compile("/handler")])
         dispatcher.register_next_step_handler(
@@ -205,7 +205,7 @@ class TestDispatcherAcceptableHandlers:
         def handler(*_):
             testing_array.append(True)
 
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
 
         dispatcher.add_handler(collector.handlers[re.compile("/handler")])
 
@@ -224,7 +224,7 @@ class TestDispatchersLifespan:
 
     @pytest.mark.asyncio
     async def test_async_dispatcher_lifespan_events(self):
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
         await dispatcher.shutdown()
 
@@ -388,7 +388,7 @@ class TestSyncDispatcherHandlersExecution:
 class TestAsyncDispatcherHandlersExecution:
     @pytest.mark.asyncio
     async def test_raising_exception_when_missing_handler_call(self, message_data):
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
 
         with pytest.raises(BotXException):
@@ -406,7 +406,7 @@ class TestAsyncDispatcherHandlersExecution:
         async def handler(*_):
             testing_array.append("text")
 
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
         dispatcher.add_handler(collector.handlers[re.compile("/handler")])
         await dispatcher.execute_command(message_data(command="/handler"))
@@ -416,7 +416,7 @@ class TestAsyncDispatcherHandlersExecution:
 
     @pytest.mark.asyncio
     async def test_next_step_handlers_registration(self, message_data):
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
 
         collector = HandlersCollector()
@@ -448,7 +448,7 @@ class TestAsyncDispatcherHandlersExecution:
 
     @pytest.mark.asyncio
     async def test_passing_extra_arguments_into_next_step_handlers(self, message_data):
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
 
         collector = HandlersCollector()
@@ -484,7 +484,7 @@ class TestAsyncDispatcherHandlersExecution:
 
     @pytest.mark.asyncio
     async def test_allowing_extra_arguments_for_common_handlers(self, message_data):
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
 
         collector = HandlersCollector()
@@ -514,7 +514,7 @@ class TestAsyncDispatcherHandlersExecution:
 
     @pytest.mark.asyncio
     async def test_calling_common_handlers_from_next_step_handler(self, message_data):
-        dispatcher = AsyncDispatcher()
+        dispatcher = AsyncDispatcher(tasks_limit=10)
         await dispatcher.start()
         msg = message_data(command="/handler")